% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI_cal.R
\name{BMI_cal}
\alias{BMI_cal}
\alias{weight_cal}
\title{Calculate BMI}
\usage{
BMI_cal(
  height,
  weight,
  data = NULL,
  digits = 1,
  h.unit = c("m", "cm"),
  w.unit = c("kg", "g")
)

weight_cal(height, BMI = 22, data = NULL, digits = 1, h.unit = c("m", "cm"))
}
\arguments{
\item{height}{The column name or number containing the height value, if a vector or dataframe of height values is specified.}

\item{weight}{The column name or number containing the weight value, if a vector or dataframe of height values is specified.}

\item{data}{data frame.}

\item{digits}{Significant digits of the value to be output. The number of decimal places to display. If you want to display all calculated values without significant digits, specify NULL.}

\item{h.unit}{Select the unit of height in meters or centimeters.}

\item{w.unit}{Select the unit of weight in kg or g.}

\item{BMI}{numeric value. BMI, want to find weight.}
}
\description{
Calculate Body mass index (BMI) based on height and weight.
Determine weight from height and BMI
}
\details{
Height and weight values have to be numerical.
BMI is calculated by the following formula.
\deqn{BMI = weight(kg) / height(m) ^ 2}
Height and weight must be a single value or a vector of equal length.
If one is a single value and the other is a vector, the single value will be recycled and used.
}
\examples{
BMI_cal(1.75, 65)
height <- c(165, 170, 175)
weight <- c(60, 65, 70)
BMI_cal(height, weight, h.unit = "cm")
}
