% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGFR_cal.R
\name{eGFR_cal}
\alias{eGFR_cal}
\title{Calculate eGFR}
\usage{
eGFR_cal(
  value,
  age,
  sex,
  data = NULL,
  digits = 1,
  male = "male",
  female = "female",
  v.type = c("Cr", "Cys-C")
)
}
\arguments{
\item{value}{The column name or number containing the value of Creatinine or Cyc-C, if a vector or dataframe of values is specified.}

\item{age}{The column name or number containing the age value, if a vector or dataframe of age values is specified.}

\item{sex}{The column name or value of sex, if a vector or dataframe of sex values is specified.}

\item{data}{data frame.}

\item{digits}{Significant digits of the value to be output. The number of decimal places to display. If you want to display all calculated values without significant digits, specify NULL.}

\item{male}{The value that specifies the gender as male. The default value is "man".}

\item{female}{The value that specifies the gender as female. The default value is "woman".}

\item{v.type}{Select whether the value being entered is Creatinine or Cyc-C.}
}
\description{
Calculate estimation of glomerular filtration rate (eGFR) based on value of Creatinine or Cyc-C and age, sex.
}
\details{
Values of Creatinine or Cyc-c and Age values have to be numerical.
If one or two is a single value and the other is a vector, the single value will be recycled and used.
}
\note{
The calculation of eGFR is using the formula for Japanese.
}
\examples{
eGFR_cal(1, 20, "man")
df <- data.frame(value = c(1, 1),
                 age = c(20, 80),
                 sex = c("M", "F"))
eGFR_cal("value", "age", "sex", df, male = "M", female = "F")
}
\references{
Matsuo S, Imai E, Horio M, Yasuda Y, Tomita K, Nitta K, Yamagata K, Tomino Y, Yokoyama H, Hishida A; Collaborators developing the Japanese equation for estimated GFR. Revised equations for estimated GFR from serum creatinine in Japan. Am J Kidney Dis. 2009 Jun;53(6):982-92. doi: 10.1053/j.ajkd.2008.12.034. Epub 2009 Apr 1. PMID: 19339088.

Horio M, Imai E, Yasuda Y, Watanabe T, Matsuo S; Collaborators Developing the Japanese Equation for Estimated GFR. GFR estimation using standardized serum cystatin C in Japan. Am J Kidney Dis. 2013 Feb;61(2):197-203. doi: 10.1053/j.ajkd.2012.07.007. Epub 2012 Aug 11. PMID: 22892396.
}
