% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KtV_cal.R
\name{KtV_cal}
\alias{KtV_cal}
\title{Calculate Kt/V}
\usage{
KtV_cal(pre.BUN, post.BUN, time, post.weight, volume, data = NULL, digits = 2)
}
\arguments{
\item{pre.BUN}{a (non-empty) numeric vector of data values. pre dialysis BUN (mg/dL).}

\item{post.BUN}{a (non-empty) numeric vector of data values. post dialysis BUN (mg/dL).}

\item{time}{a (non-empty) numeric vector of data values. Duration of dialysis (hours).}

\item{post.weight}{a (non-empty) numeric vector of data values. Ultrafiltration volume (L).}

\item{volume}{a (non-empty) numeric vector of data values. post dialysis weight (kg).}

\item{data}{data frame.}

\item{digits}{integer indicating the number of decimal places. If you want to display all calculated values without significant digits, specify NULL.}
}
\description{
Calculate variable-volume single-pool (VVSP) Kt/V from BUN before and after dialysis, dialysis time, post-dialysis weight, and ultrafiltration Volume.
}
\details{
Daugirdas' formula is used to calculate Kt/V.
}
\references{
Daugirdas JT. Second generation logarithmic estimates of single-pool variable volume Kt/V: an analysis of error. J Am Soc Nephrol. 1993 Nov;4(5):1205-13. doi: 10.1681/ASN.V451205. PMID: 8305648.
}
